+++ HARUCHI/HARUCHI/App/Home/HomeViewModel.swift
@@ -17,9 +17,30 @@ class HomeViewModel: ObservableObject {
     @Published var navigateToReceipt: Bool = false
     @Published var navigateToHomeMain: Bool = false
     @Published var showMainButton: Bool = false
+    @Published var weekData: [WeekCalendarModel] = []
+    
+    
+    init() {
+        setupDummyData()
+    }
     
     
     func hideKeyboard() {
         UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+    }
+    
+    func setupDummyData() {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd"
+
+        weekData = [
+            WeekCalendarModel(date: formatter.date(from: "2024-08-05")!, icon: "calendar_smiling", value: 0),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-06")!, icon: "calendar_pouting", value: 0),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-07")!, icon: "calendar_smiling", value: 0),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-08")!, icon: "calendar_base", value: 20000),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-09")!, icon: "calendar_base", value: 20000),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-10")!, icon: "calendar_base", value: 20000),
+            WeekCalendarModel(date: formatter.date(from: "2024-08-11")!, icon: "calendar_base", value: 20000),
+        ]
     }
 }
