--- HARUCHI/HARUCHI/Components/WeekCalendar/WeekCalendarModel.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-struct WeekCalendarModel {
-    var date: Date
-    var icon: String // 이미지 이름
-    var value: Int? // 해당 날짜의 값
-}
